requirements.txt
langchain==0.1.0
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.3.2
python-docx==0.8.11
matplotlib==3.8.2
seaborn==0.13.0
tiktoken==0.5.1 
langchain-community